fix : useDebounce적용
