+++ src/basic/utils/hooks/useDebounce.ts
@@ -1,0 +1,22 @@
+import { useState, useEffect } from "react";
+
+/**
+ * @param value 변경될 값
+ * @param delay 디바운스 지연 시간 (기본값: 500ms)
+ * @returns 디바운스된 값
+ */
+export function useDebounce<T>(value: T, delay: number = 500): T {
+  const [debouncedValue, setDebouncedValue] = useState<T>(value);
+
+  useEffect(() => {
+    const handler = setTimeout(() => {
+      setDebouncedValue(value);
+    }, delay);
+
+    return () => {
+      clearTimeout(handler);
+    };
+  }, [value, delay]);
+
+  return debouncedValue;
+}
