+++ src/refactoring(hint)/utils/hooks/useDebounce.ts
@@ -6,23 +6,6 @@
 //
 // 사용 예시: 검색어 입력 디바운싱
 
-import { useEffect, useState } from "react";
-
-// export function useDebounce<T>(value: T, delay: number): T {
-//   // TODO: 구현
-
-// }
 export function useDebounce<T>(value: T, delay: number): T {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
+  // TODO: 구현
 }
